reorganização dos serviços e criação do service para produtos

+++ LojaVendeTudo.API/Models/Produto.cs
@@ -7,13 +7,13 @@ namespace LojaVendeTudo.API.Models
     public class Produto : AcessoBanco
     {
         [IdSequence]
-        public long ProdutoID { get; set; }
+        public int ProdutoID { get; set; }
         public string Descricao  { get; set; }
         public string Nome { get; set; }
 	    public string Marca { get; set; }
 	    public string Fabricante { get; set; }
 	    public string Peso { get; set; }
 	    public decimal PrecoUnitario { get; set; }
-	    public long fk_Fornecedor { get; set; }
+	    public int fk_Fornecedor { get; set; }
     }
 }
